+++ src/main/java/nextstep/subway/application/StationService.java
@@ -1,5 +1,9 @@
-package nextstep.subway.station;
+package nextstep.subway.application;
 
+import nextstep.subway.application.dto.StationRequest;
+import nextstep.subway.application.dto.StationResponse;
+import nextstep.subway.domain.Station;
+import nextstep.subway.domain.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
