+++ src/main/java/nextstep/subway/application/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section;
+package nextstep.subway.application.dto;
 
 public class SectionRequest {
     private final Long upStationId;
