+++ src/main/java/nextstep/subway/domain/LineRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
