+++ src/main/java/nextstep/subway/domain/SectionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section;
+package nextstep.subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
