+++ src/main/java/nextstep/subway/domain/Station.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station;
+package nextstep.subway.domain;
 
 import javax.persistence.*;
 import java.util.Objects;
