+++ src/main/java/nextstep/subway/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station;
+package nextstep.subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
