+++ src/test/java/nextstep/subway/unit/LineServiceMockTest.java
@@ -1,7 +1,7 @@
 package nextstep.subway.unit;
 
-import nextstep.subway.line.LineRepository;
-import nextstep.subway.station.StationService;
+import nextstep.subway.domain.LineRepository;
+import nextstep.subway.application.StationService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
